capstone-projects/kltn-finance/kltn-fm-storage/face_id_service/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pydantic
mediapipe
opencv-python
numpy
pymongo
python-dotenv
langchain
langchain-core
langchain-community
faiss-cpu
pypdf
sentence-transformers
langchain-huggingface
huggingface_hub